requirements.txt
groq>=0.4.0
requests>=2.31.0
edge-tts>=6.1.0
python-dotenv>=1.0.0
Pillow>=10.0.0
imageio-ffmpeg>=0.4.0
google-api-python-client>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.2.0
